Rename model-editor-queries.ts to model-editor-queries-setup.ts
