+++ extensions/ql-vscode/src/model-editor/model-editor-queries-setup.ts
